make build1-4 (Java 1.4 distribution) compile again

+++ prov/src/main/java/org/bouncycastle/jcajce/provider/keystore/pkcs12/PKCS12Util.java
@@ -78,7 +78,7 @@ class PKCS12Util
      */
     static int getKeySize(AlgorithmIdentifier algorithmIdentifier)
     {
-        Integer keySize = KEY_SIZES.get(algorithmIdentifier.getAlgorithm());
+        Integer keySize = (Integer)KEY_SIZES.get(algorithmIdentifier.getAlgorithm());
         return keySize == null ? -1 : keySize.intValue();
     }
